usm: http2: tests: move utilities of http2 to the new file
